requirements.txt
Django                    3.0.2
django-cors-headers       3.2.1
django-filter             2.2.0
djangorestframework       3.11.0
GitPython                 3.0.5
Markdown                  3.1.1
mysqlclient               1.4.6
paramiko                  2.7.1
tornado                   6.0.3
djangorestframework-jwt	  1.11.0
django-celery             3.3.1
flower                    0.9.3


